Coverity 371993: Add explicit casts to silence coverity warning.
Coverity is unhappy that we do (int/2) and then use it in a double
context, worried that we might have inadvertantly thrown away any
fraction. We are deliberately throwing away the fraction, so hopefully
the explicit cast will satisfy it.

+++ base/gxshade1.c
@@ -196,8 +196,8 @@ typedef struct
  */
 
 
-#define MIN_CLIP_LIMIT (fixed2int(min_fixed)/2)
-#define MAX_CLIP_LIMIT (fixed2int(max_fixed)/2)
+#define MIN_CLIP_LIMIT ((int)(fixed2int(min_fixed)/2))
+#define MAX_CLIP_LIMIT ((int)(fixed2int(max_fixed)/2))
 
 static int not_clipped_away(const gs_point *p)
 {
